requirements.txt
scikit-image==0.18.2
streamlit==0.86.0
tensorflow==2.5.0
Pillow==8.3.1
numpy

